Change import dependency graph
get_service function is now defined in config
